chore: rename socket.js to server.js
