+++ app/bin/www
@@ -25,7 +25,7 @@ const server = http.createServer(app);
  * Socket.io
  */
 
-const { io } = require('../socket/socket.js');
+const { io } = require('../socket/server.js');
 
 io.attach(server);
